提交

+++ packages/server/src/modules/stock/stock.service.ts
@@ -1,17 +1,30 @@
 import { InjectRepository } from '@nestjs/typeorm';
-import { FindOptionsWhere, Repository, DataSource } from 'typeorm';
+import { FindOptionsWhere, Repository, DataSource, In } from 'typeorm';
 import * as validator from 'class-validator';
 import { plainToInstance } from 'class-transformer';
-import { WorkSheet, utils, write } from 'xlsx';
+import { WorkBook, WorkSheet, utils, write } from 'xlsx';
 import * as fs from 'fs';
 import * as moment from 'moment';
 import { StockEntity } from './stock.entity';
-import { SearchDto, StockParseDto } from './stock.dto';
+import {
+  SearchDto,
+  StockInnerParseResult,
+  StockParseDto,
+  StockSaveDto,
+  Type,
+} from './stock.dto';
 import { ErrorConstant } from 'src/constant/error';
-import { stockHeaderMap, stockSheetName, tmpPath } from '../../constant/file';
+import {
+  stockHeaderMap,
+  stockSheetName,
+  tmpPath,
+  dateFormat,
+} from '../../constant/file';
 import { ProductService } from '../product/product.service';
 import { StoreService } from '../store/store.service';
+import { ConfigService } from '../config/config.service';
 import { getTree } from './util';
+import { fixImportedDate } from '../../utils';
 import { appLogger } from 'src/logger';
 
 export class StockService {
@@ -21,6 +34,7 @@ export class StockService {
     private dataSource: DataSource,
     private productService: ProductService,
     private storeService: StoreService,
+    private configService: ConfigService,
   ) {}
 
   /**
@@ -71,9 +85,12 @@ export class StockService {
    */
   async findAll(query: SearchDto): Promise<StockEntity[]> {
     const where: FindOptionsWhere<StockEntity> = {};
-    const { week, customerId } = query;
+    const { week, weeks, customerId } = query;
     if (validator.isNotEmpty(week)) {
       where.week = week;
+    }
+    if (validator.isNotEmpty(weeks)) {
+      where.week = In(weeks);
     }
     if (validator.isNotEmpty(customerId)) {
       where.customer = {
@@ -90,9 +107,12 @@ export class StockService {
    */
   async findCount(query: SearchDto): Promise<number> {
     const where: FindOptionsWhere<StockEntity> = {};
-    const { week, customerId } = query;
+    const { week, weeks, customerId } = query;
     if (validator.isNotEmpty(week)) {
       where.week = week;
+    }
+    if (validator.isNotEmpty(weeks)) {
+      where.week = In(weeks);
     }
     if (validator.isNotEmpty(customerId)) {
       where.customer = {
@@ -108,14 +128,18 @@ export class StockService {
    * 解析数据插入数据
    */
   async parseSheet(
+    workbook: WorkBook,
     sheet: WorkSheet,
     fileName: string,
     body: StockParseDto,
     creatorId: number,
-  ): Promise<number | ErrorConstant> {
+  ): Promise<StockInnerParseResult | ErrorConstant> {
     const { weekStartDate, weekEndDate, week, customerId } = body;
 
-    const arrs = utils.sheet_to_json<any[]>(sheet, { header: 1 });
+    const arrs = utils.sheet_to_json<any[]>(sheet, {
+      header: 1,
+      dateNF: dateFormat,
+    });
     if (arrs.length === 0) {
       return;
     }
@@ -130,7 +154,31 @@ export class StockService {
       }
     });
 
-    const data = utils.sheet_to_json(sheet);
+    // 类型1为基于周进行导入，类型2为基于excel中的日期进行导入，类型2需要根据“日期”字段计算出周
+    let importType = 1;
+    // week为空，说明为类型2
+    if (!week || colMap.hasOwnProperty('date')) {
+      importType = 2;
+    }
+
+    // 查询周开始日
+    let weekStartIndex = '1';
+    if (importType === 2) {
+      weekStartIndex = await this.configService.get('getWeekStartIndex');
+      if (!weekStartIndex) {
+        weekStartIndex = '1';
+      }
+      moment.locale('zh-cn', {
+        week: {
+          dow: Number(weekStartIndex),
+        },
+      });
+    }
+
+    const data = utils.sheet_to_json(sheet, {
+      dateNF: dateFormat,
+    });
+    const is_date1904 = workbook.Workbook.WBProps.date1904;
     const result = data.map((item) => {
       const temp: Partial<StockEntity> = {
         weekStartDate,
@@ -144,6 +192,10 @@ export class StockService {
         if (item.hasOwnProperty(oldKey)) {
           const newKey = stockHeaderMap[oldKey];
           temp[newKey] = item[oldKey];
+          // 设置sheetjs date格式的时差，https://github.com/SheetJS/sheetjs/issues/1565
+          if (temp[newKey] instanceof Date) {
+            temp[newKey] = fixImportedDate(temp[newKey], is_date1904);
+          }
         }
       }
       return temp;
@@ -161,7 +213,7 @@ export class StockService {
     for (let rowIndex = 0; rowIndex < entities.length; rowIndex++) {
       const errorsTemp = [];
       const entity = entities[rowIndex];
-      const { productName, storeName, quantity, price, total } = entity;
+      const { productName, storeName, quantity, price, total, date } = entity;
 
       // 产品名称不在系统内，或者没填写，必填字段
       if (!productName || !allProductNames.includes(productName)) {
@@ -218,6 +270,44 @@ export class StockService {
         errorsTemp.push(errMsg);
       }
 
+      // 时间不是日期类型
+      if (date && !validator.isDate(date)) {
+        // 单元格位置文本，A1 B2
+        const position = `${utils.encode_cell({
+          c: colMap.date,
+          r: rowIndex + 1,
+        })}`;
+        const errMsg = `位置: ${position} 时间"${date}"不是日期类型`;
+        errorsTemp.push(errMsg);
+      }
+
+      // 时间不是日期类型
+      if (importType === 2 && !date) {
+        // 单元格位置文本，A1 B2
+        const position = `${utils.encode_cell({
+          c: colMap.date,
+          r: rowIndex + 1,
+        })}`;
+        const errMsg = `位置: ${position} 有"库存 - 时间"列的时候"库存 - 时间"必填`;
+        errorsTemp.push(errMsg);
+      }
+
+      // 类型2情况下为week、weekStartDate、weekEndDate添加值
+      if (importType === 2 && date) {
+        const dateMoment = moment(date);
+        const dateStr = dateMoment.format(dateFormat);
+        const weekStr = dateMoment.format('gggg-w');
+        const startDate = dateMoment.startOf('week').format(dateFormat);
+        const endDate = dateMoment.endOf('week').format(dateFormat);
+        entity.week = weekStr;
+        // @ts-ignore
+        entity.date = dateStr;
+        // @ts-ignore
+        entity.weekStartDate = startDate;
+        // @ts-ignore
+        entity.weekEndDate = endDate;
+      }
+
       // 有失败的话，就在entity中添加失败原因，方便进行导出
       if (errorsTemp.length > 0) {
         data[rowIndex]['失败原因'] = errorsTemp.join('，');
@@ -250,6 +340,29 @@ export class StockService {
       );
       return e;
     }
+    let weeks = entities.map((item) => item.week).filter((item) => !!item);
+    weeks = [...new Set(weeks)];
+    const repeatData = await this.findAll({
+      customerId,
+      weeks,
+    });
+    let repeatWeeks = repeatData
+      .map((item) => item.week)
+      .filter((item) => !!item);
+    repeatWeeks = [...new Set(repeatWeeks)];
+    const repeatWeekCount = repeatWeeks.length;
+
+    return {
+      data: entities,
+      repeatWeekCount,
+      repeatWeeks,
+    };
+  }
+
+  async save({ data, customerId, type }: StockSaveDto) {
+    const entities = plainToInstance(StockEntity, data);
+    let weeks = entities.map((item) => item.week).filter((item) => !!item);
+    weeks = [...new Set(weeks)];
 
     // 增加一个事务，先删除，再保存
     let res;
@@ -257,13 +370,16 @@ export class StockService {
     await queryRunner.connect();
     await queryRunner.startTransaction();
     try {
-      await queryRunner.manager
-        .createQueryBuilder()
-        .delete()
-        .from(StockEntity)
-        .where('week = :week', { week })
-        .andWhere('customer_id = :customerId', { customerId })
-        .execute();
+      // 覆盖
+      if (type === Type.cover) {
+        await queryRunner.manager
+          .createQueryBuilder()
+          .delete()
+          .from(StockEntity)
+          .where('week IN (:...weeks)', { weeks })
+          .andWhere('customer_id = :customerId', { customerId })
+          .execute();
+      }
       res = await queryRunner.manager.save(entities);
       await queryRunner.commitTransaction();
     } catch (err) {
